Rename hello to main.
