+++ .travis.yml
@@ -3,4 +3,4 @@ python:
   - "2.6"
   - "2.7"
 install: "pip install -r requirements.txt"
-script: python test_hello.py
+script: python test_main.py
